Change repo structure
