+++ pkg/flaki/logging_test.go
@@ -1,0 +1,67 @@
+package flaki
+
+import (
+	"context"
+	"math/rand"
+	"strconv"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLoggingMiddleware(t *testing.T) {
+	var mockLogger = &mockLogger{}
+	var mockFlaki = &mockFlaki{}
+
+	// Context with correlation ID.
+	rand.Seed(time.Now().UnixNano())
+	var id = strconv.FormatUint(rand.Uint64(), 10)
+	var ctx = context.WithValue(context.Background(), "correlation_id", id)
+
+	var srv = New(mockFlaki)
+	srv = MakeLoggingMiddleware(mockLogger)(srv)
+
+	// NextID.
+	mockLogger.Called = false
+	mockLogger.CorrelationID = ""
+	srv.NextID(ctx)
+	assert.True(t, mockLogger.Called)
+	assert.Equal(t, id, mockLogger.CorrelationID)
+
+	// NextValidID.
+	mockLogger.Called = false
+	mockLogger.CorrelationID = ""
+	srv.NextValidID(ctx)
+	assert.True(t, mockLogger.Called)
+	assert.Equal(t, id, mockLogger.CorrelationID)
+
+	// NextID without correlation ID.
+	var f = func() {
+		srv.NextID(context.Background())
+	}
+	assert.Panics(t, f)
+
+	// NextValidID without correlation ID.
+	f = func() {
+		srv.NextValidID(context.Background())
+	}
+	assert.Panics(t, f)
+}
+
+// Mock Logger.
+type mockLogger struct {
+	Called        bool
+	CorrelationID string
+}
+
+func (l *mockLogger) Log(keyvals ...interface{}) error {
+	l.Called = true
+
+	for i, kv := range keyvals {
+		if kv == "correlation_id" {
+			l.CorrelationID = keyvals[i+1].(string)
+		}
+	}
+	return nil
+}
