--- pkg/flaki/transport/grpc/grpc_test.go
@@ -1,1 +1,0 @@
-package flakigrpc
