--- pkg/flaki/transport/http/http_test.go
@@ -1,1 +1,0 @@
-package flakihttp
